+++ pkg/manifests/manifests.go
@@ -4228,6 +4228,7 @@ func addRemoteWriteConfigs(clusterID string, rw []monv1.RemoteWriteSpec, rwTarge
 			BearerTokenFile:     target.BearerTokenFile,
 			ProxyURL:            target.ProxyURL,
 			MetadataConfig:      target.MetadataConfig,
+			OAuth2:              target.OAuth2,
 		}
 		if target.TLSConfig != nil {
 			rwConf.TLSConfig = &monv1.TLSConfig{